Renamed Nofitications related files
