+++ library/Notifications/Mobile/System.php
@@ -1,15 +1,14 @@
 <?php
 
-namespace Gewaer\Notifications\PushNotifications;
+namespace Gewaer\Notifications\Mobile;
 
 use Namshi\Notificator\Notification;
 use Gewaer\Contracts\PushNotifications as PushNotificationsContract;
-use Gewaer\Models\Users;
 use Gewaer\Models\Notifications;
-use Gewaer\Notifications\PushNotifications\PushNotifications as PushNotifications;
+use Gewaer\Notifications\Mobile\Mobile;
 use Gewaer\Traits\NotificationsTrait;
 
-class AppsPushNotifications extends PushNotifications implements PushNotificationsContract
+class System extends Mobile implements PushNotificationsContract
 {
     /**
      * Notifications Trait
@@ -25,8 +24,8 @@ class AppsPushNotifications extends PushNotifications implements PushNotificatio
         /**
          * Create a new database record
          */
-        self::create($this->user, $this->content, Notifications::APPS, $this->systemModule);
+        self::create($this->user, $this->content, Notifications::SYSTEM, $this->systemModule);
 
-        return $this->content . " From Apps";
+        return $this->content . " From System";
     }
 }
